Rename patch files
